Move jctest to ext folder in test Remove all submodules (now unused)
